resolves #22: Fixes spec to match instructions
Multiple students have asked questions about this and there's 2 issues filed already. It seems like a simple fix... The instructions say to create a migration to change the data type of this column to a string. But the tests are expecting it to be a boolean. Thus, following the instructions makes this test fail.

This fix brings the instructions and the specs into agreement.